+++ CHANGELOG.md
@@ -4,6 +4,10 @@ FEATURES:
 
 * **New Resource**: `tencentcloud_kubernetes_node_pool` to support node management.
 
+DEPRECATED:
+
+* Resource: `tencentcloud_kubernetes_as_scaling_group` replaced by `tencentcloud_kubernetes_node_pool`.
+
 
 ## 1.51.1 (December 22, 2020)